+++ config/database.go
@@ -7,7 +7,7 @@ import (
 )
 
 func GetConexao() *sql.DB {
-	db, err := sql.Open("mysql", "root:159357@tcp(127.0.0.1:3306)/db_encurtador_url_go")
+	db, err := sql.Open("mysql", "root:@tcp(127.0.0.1:3306)/db_encurtador_url_go")
 
 	if err != nil {
 		panic(err.Error())
