+++ main.go
@@ -2,20 +2,16 @@ package main
 
 import (
 	"encoding/json"
-	"fmt"
 	"log"
 	"math/rand"
 	"net/http"
 
 	"github.com/gorilla/mux"
-	"github.com/ropehapi/encurtador-url-go/config"
+	"github.com/ropehapi/encurtador-url-go/model"
 )
 
-type Relation struct {
-	Url, Code string
-}
-
 var letterRunes = []rune("abcdefghijklmnopqrstuvwxyz")
+
 func RandStringRunes(n int) string {
 	b := make([]rune, n)
 	for i := range b {
@@ -33,41 +29,25 @@ func main() {
 	log.Fatal(http.ListenAndServe(":8000", r))
 }
 
+// TODO: I must validate if the code already exists in a relation
 func Encurta(w http.ResponseWriter, r *http.Request) {
 	url := r.FormValue("url")
 	code := RandStringRunes(6)
-	
-	db := config.GetConexao()
-	defer db.Close()
 
-	//I should test if this sentence really get the existing code without any treatment in the result
-	existingCode, err := db.Query(fmt.Sprintf("SELECT * FROM relation WHERE code = '%s'", code))
+	relation := model.Relation{
+		Url:  url,
+		Code: code,
+	}
+
+	err := relation.Store()
 	if err != nil {
-		w.WriteHeader(http.StatusBadRequest)
-		json.NewEncoder(w).Encode(err.Error())
+		panic(err.Error())
 	}
 
-	//Must test if this first case calls the method itself
-	if existingCode != nil {
-		Encurta(w, r)
-	}else{
-		relation := Relation{
-			Url:  url,
-			Code: code,
-		}
-	
-		_, err = db.Exec(fmt.Sprintf("INSERT INTO relation (url, code) VALUES ('%s','%s')", relation.Url, relation.Code))
-		if err != nil {
-			w.WriteHeader(http.StatusBadRequest)
-			json.NewEncoder(w).Encode(err.Error())
-		}
-	
-	
-		w.WriteHeader(http.StatusOK)
-		json.NewEncoder(w).Encode(relation.Code)
-	}
+	w.WriteHeader(http.StatusOK)
+	json.NewEncoder(w).Encode(relation.Code)
 }
 
 func Desencurta(w http.ResponseWriter, r *http.Request) {
-	
+
 }
